Move get_size_bytes to lmcache.utils
It is engine-agnostic and the core (memory management, allocators)
imported it from the vLLM integration module, inverting the layering
and creating an import cycle through lmcache.v1.gpu_connector. The
old name re-exports; the cycle-dodging function-level import of
KV_LAYOUT_NAMES becomes a plain module import.
