+++ lmcache/v1/memory_allocators/ad_hoc_memory_allocator.py
@@ -8,8 +8,7 @@ import torch
 
 # First Party
 from lmcache import torch_dev
-from lmcache.integration.vllm.utils import get_size_bytes
-from lmcache.utils import _lmcache_nvtx_annotate
+from lmcache.utils import _lmcache_nvtx_annotate, get_size_bytes
 from lmcache.v1.memory_management import (
     MemoryAllocatorInterface,
     MemoryFormat,
